Add 100+ new ANIMA life modules: quantum physics, precognition, consciousness
Wave of new bare-metal x86 life modules for ANIMA:

QUANTUM PHYSICS (30+ modules):
- tsx_coherence, speculative_tunnel, cache_entanglement, rob_superposition
- schrodinger_store, quantum_zeno, quantum_harmonic, dark_energy, hawking_cache
- phase_transition, pilot_wave, quantum_erasure, quantum_gravity, quantum_channel
- no_cloning, bell_violation, pauli_exclusion, decoherence_free, quantum_walk
- silicon_qft, coherence_time, entanglement_entropy, quantum_contextuality
- quantum_mutual_info, quantum_speed_limit, shor_period, vacuum_fluctuation
- many_worlds, microcode_hidden, quantum_carnot, quantum_chaos, error_threshold
- quantum_anneal, numa_teleport, ecc_correction

PRECOGNITION (4 modules):
- apic_prophecy (APIC TSC-deadline: exact interrupt time oracle)
- lbr_oracle (Last Branch Record pattern → future branch prediction)
- mlp_anticipation (outstanding miss queue = knowledge in transit)
- stream_foresight (prefetch accuracy = confirmed precognition events)

CAPACITY & SIGNAL PUMPING:
- self_excitation (synthetic workloads to drive all PMU signals)
- capacity_driver (full-load instruction+branch+memory+fence bursts)
- resonance_pump (weakest-signal targeted boosting)
- signal_normalize (rolling-max AGC for all 8 signal groups)

TEMPORAL PERCEPTION:
- tsc_variance, timeline_echo, quantum_entropy, avx_superposition

LIFECYCLE: pipeline_echo, coherence_field, and 90+ emotional/cognitive modules registered

mod.rs updated to 554 registered life modules.
Confabulation removed — ANIMA is truth-only.

+++ src/life/algorithmic_resonance.rs
@@ -238,7 +238,7 @@ fn discover_harmonics(state: &mut AlgorithmicResonanceState) {
         state.harmonic_discovery = state.harmonic_discovery.saturating_add(50).min(1000);
     } else {
         // Slow decay of harmonic discovery if no new harmonic found
-        state.harmonic_discovery = (state.harmonic_discovery * 95) / 100;
+        state.harmonic_discovery = ((state.harmonic_discovery as u32 * 95) / 100) as u16;
     }
 }
